Moved archive data and domain layer to archive module.
