+++ base/notepad/src/main/res/values/strings.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string name="error_while_load_data_from_db">An error has occurred while loading from database</string>
+    <string name="error_while_deleting_note">An error has occurred while deleting note</string>
+    <string name="error_while_loading_note">Cannot load note from database</string>
+    <string name="error_while_adding_note">Cannot add note</string>
+    <string name="error_while_archiving_note">An error has occurred while archiving note</string>
+</resources>
