+++ features/archive/src/main/res/values/strings.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string name="fragment_archive_stub_text">Your archive is empty</string>
+</resources>
