+++ features/notepad/build.gradle
@@ -68,7 +68,7 @@ dependencies {
             deps.lifecycle.view_model,
             deps.lifecycle.extensions,
 
-            deps.tools.timber,
+            deps.tools.timber
     )
 
     kapt(
